% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adfi_get.R
\name{adfi_get}
\alias{adfi_get}
\title{ADFI get from pig performance test station csv data}
\usage{
adfi_get(data, adg_res)
}
\arguments{
\item{data}{A data frame or data table containing the nedap or fire pig performance test data to be processed. Columns must include 'visit_time', 'location', 'responder', 'feed_intake'.}

\item{adg_res}{ADG results from function of adg_get().}
}
\value{
A list containing:
\itemize{
\item dfi_info: A data.table containing DFI statistics
\item dfi_data: A data.table containing processed sample data
}
}
\description{
ADFI get from pig performance test station csv data
}
\examples{
nedap_csv_data <- mintyr::nedap
adg_results <- adg_get(data = nedap_csv_data)
adfi_results <- adfi_get(data = nedap_csv_data, adg_res = adg_results)
head(adfi_results$adfi_info)
}
