% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adg_get.R
\name{adg_get}
\alias{adg_get}
\title{ADG get from pig performance test station csv data}
\usage{
adg_get(
  data,
  my_break = NULL,
  range_offset = 0.5,
  threshold = 1,
  save_path = NULL
)
}
\arguments{
\item{data}{A data frame or data table containing the nedap or fire pig performance test data to be processed. Columns must include 'visit_time', 'location', 'responder', 'feed_intake'.}

\item{my_break}{Optional, a numeric vector of length 2, indicates target weight range for calculating ADG, default NULL. If not NULL, ADG will be calculated within this range.}

\item{range_offset}{Optional, a numeric value, default 0.5.
Used to extend the target weight range specified in my_break to avoid border effect.
For example, if my_break is c(60,90) and range_offset is 0.5, the actual range for analysis will be 57.5~92.5.}

\item{threshold}{Optional, a numeric value, default 1, used as the threshold to identify outliers in RANSAC regression, usually 0~2.}

\item{save_path}{Optional, a character string specifying where to save the generated growth curve images. If not NULL, images will be saved to this path.}
}
\value{
A list containing:
\itemize{
\item adg_info: A data.table containing ADG statistics
\item adg_data: A data.table containing processed sample data
}
}
\description{
ADG get from pig performance test station csv data
}
\examples{
list_csv <- list.files("C:/Users/Dell/Downloads/location301", pattern = ".csv", full.names = TRUE)
data <- import_csv(list_csv)
adg_res <- adg_get(data = data)
}
