% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfi_get.R
\name{dfi_get}
\alias{dfi_get}
\title{ADFI get from pig performance test station csv data}
\usage{
dfi_get(data, my_break = NULL, adg_about)
}
\arguments{
\item{data}{A data frame or data table containing the nedap or fire pig performance test data to be processed. Columns must include 'visit_time', 'location', 'responder', 'feed_intake'.}

\item{my_break}{Optional, a numeric vector of length 2, indicates target weight range for calculating DFI, default NULL. If not NULL, DFI will be calculated within this range.}

\item{adg_about}{ADG results from function of adg_get().}
}
\value{
A list containing:
\itemize{
\item dfi_info: A data.table containing DFI statistics
\item dfi_data: A data.table containing processed sample data
\item feedintake_order: A data.table containing dfi order data
}
}
\description{
ADFI get from pig performance test station csv data
}
\examples{
list_csv <- list.files("C:/Users/Dell/Downloads/location301", pattern = ".csv", full.names = TRUE)
data <- import_csv(list_csv)
adg_res <- adg_get(data = data)
dfi_get <- dfi_get(data = data, adg_about = adg_res)
}
