% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_get.R
\name{fcr_get}
\alias{fcr_get}
\title{FCR get from pig performance test station csv data}
\usage{
fcr_get(
  data,
  my_break = NULL,
  range_offset = 0.5,
  threshold = 1,
  save_path = NULL
)
}
\arguments{
\item{data}{A data frame or data table containing the nedap or fire pig performance test data to be processed. Columns must include 'visit_time', 'location', 'responder', 'feed_intake'.}

\item{my_break}{Optional, a numeric vector of length 2, indicates target weight range for calculating ADG, default NULL. If not NULL, ADG will be calculated within this range.}

\item{range_offset}{Optional, a numeric value, default 0.5.
Used to extend the target weight range specified in my_break to avoid border effect.
For example, if my_break is c(60,90) and range_offset is 0.5, the actual range for analysis will be 57.5~92.5.}

\item{threshold}{Optional, a numeric value, default 1, used as the threshold to identify outliers in RANSAC regression, usually 0~2.}

\item{save_path}{Optional, a character string specifying where to save the generated growth curve images. If not NULL, images will be saved to this path.}
}
\description{
FCR get from pig performance test station csv data
}
\examples{
data <- rio::import("C:/Users/Dell/Downloads/ppt_test_data.xlsx")
fcr_res <- fcr_get(data = data, my_break = c(30, 100))
}
