% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_summary.R
\name{fcr_summary}
\alias{fcr_summary}
\title{FCR summary of nedap pig performance station}
\usage{
fcr_summary(data)
}
\arguments{
\item{data}{A data.frame or a data.table. The data should include at least one pair of weight columns ("min_weight_cut" and "max_weight_cut" or "min_weight_origin" and "max_weight_origin") and the following common columns: "lm_slope", "test_days", "origin_dfi", "corrected_dfi", "fcr". The column "cor_fcr" is optional.}
}
\value{
A data.frame that includes the statistics of each column.
}
\description{
FCR summary of nedap pig performance station
}
\examples{
list_csv <- list.files("C:/Users/Dell/Downloads/location301", pattern = ".csv", full.names = TRUE)
data <- import_csv(list_csv)
fcr_res <- fcr_get(data = data, my_break = c(30, 100))
fcr_summary(fcr_res[[1]])
}
