% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_csv.R
\name{import_csv}
\alias{import_csv}
\title{Import CSV files using specified R packages}
\usage{
import_csv(file_list, package = "data.table", ...)
}
\arguments{
\item{file_list}{A list of file paths to CSV files}

\item{package}{Package to use for reading CSV. Options are "data.table", "vroom", or "readr". Default is "data.table"}

\item{...}{Additional arguments to pass to the read function}
}
\value{
A combined data.frame/data.table of all input CSV files
}
\description{
Import CSV files using specified R packages
}
\examples{
list_csv <- list.files("C:/Users/Dell/Downloads/location301", pattern = ".csv", full.names = TRUE)
data <- import_csv(list_csv)
}
