% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcr_get.R
\name{fcr_get}
\alias{fcr_get}
\title{Calculate Feed Conversion Ratio (FCR)}
\usage{
fcr_get(adg_res, adfi_res)
}
\arguments{
\item{adg_res}{A list containing ADG results, including 'adg_info' and 'adg_data' from adg_get().}

\item{adfi_res}{A list containing ADFI results, including 'adfi_info' and 'adfi_data' from adfi_get().}
}
\value{
A list containing:
\item{fcr_res}{A data.table with FCR results, including 'fcr' and 'cor_fcr' (if applicable)}
\item{adg_info}{The original ADG info data.table}
\item{adg_data}{The original ADG data}
\item{adfi_info}{The original ADFI info data.table}
\item{adfi_data}{The original ADFI data}
}
\description{
This function calculates the Feed Conversion Ratio (FCR) based on average daily gain (ADG) and average daily feed intake (ADFI) results.
It also calculates a corrected FCR if stage information is available.
}
\note{
\itemize{
\item This function assumes that 'adg_info' and 'adfi_info' are data.tables with columns 'responder' and 'location'.
\item The function expects 'corrected_dfi' and 'lm_slope' columns in the joined data for FCR calculation.
\item The corrected FCR calculation is only performed if a 'stage' column is present in the joined data.
\item The variables 'min_weight_cut' and 'max_weight_cut' should be defined in the global environment or passed as parameters for the corrected FCR calculation.
}
}
\examples{
nedap_csv_data <- mintyr::nedap
adg_results <- adg_get(data = nedap_csv_data)
adfi_results <- adfi_get(data = nedap_csv_data, adg_res = adg_results)
fcr_results <- fcr_get(adg_res = adg_results, adfi_res = adfi_results)
head(fcr_results$fcr_res)
head(fcr_results$fcr_summary)
}
